[UI] POLISHED MARKET PRODUCT CARDS
